Added In Stu's Changes

+++ CMS/CMS/src/Controller/DeleteTipCtrl.java
@@ -1,5 +1,7 @@
 package Controller;
 
+import javax.swing.JOptionPane;
+
 import Model.Model;
 import View.MainContainer;
 import View.TipsPanel;
@@ -7,6 +9,17 @@ import View.TipsPanel;
 public class DeleteTipCtrl {
 	public DeleteTipCtrl(Model model, MainContainer view) {
 		TipsPanel panel = (TipsPanel) view.getActivePanel();
-		model.deleteTip(panel.getSelectedDelete());
+		String tip = panel.getSelectedDelete();
+		int sure = JOptionPane.showConfirmDialog(null,
+				"Are you sure you wish to deleted the '" + tip + "' tip?",
+				"Are you sure?", JOptionPane.YES_NO_OPTION,
+				JOptionPane.INFORMATION_MESSAGE);
+		if (sure == JOptionPane.YES_OPTION) {
+			model.deleteTip(panel.getSelectedDelete());
+			JOptionPane.showMessageDialog(null, "Tip '" + tip
+					+ "' has been successfully deleted.", "Success",
+					JOptionPane.INFORMATION_MESSAGE);
+		}
+
 	}
 }
